Re-organized code to have better visibility
