+++ src/actions/dev.js
@@ -19,17 +19,22 @@
 
 const watch = require('./watch')
 const serve = require('./serve')
-const checkForAppPath = require('../helpers/checkForAppPath')
+const sequence = require('../helpers/sequence')
+const buildHelpers = require('../helpers/build')
 
 module.exports = () => {
-  checkForAppPath()
-  watch(serve, () => {
-    console.log('')
-    if (process.env.LNG_LIVE_RELOAD === 'true') {
-      console.log('Navigate to web browser to see the changes')
-    } else {
-      console.log('Reload your web browser to see the changes')
-    }
-    console.log('')
-  })
+  return sequence([
+    () => buildHelpers.ensureLightningApp(),
+    () => {
+      watch(serve, () => {
+        console.log('')
+        if (process.env.LNG_LIVE_RELOAD === 'true') {
+          console.log('Navigate to web browser to see the changes')
+        } else {
+          console.log('Reload your web browser to see the changes')
+        }
+        console.log('')
+      })
+    },
+  ])
 }
