+++ src/actions/dist.js
@@ -3,17 +3,16 @@ const fs = require('fs')
 const sequence = require('../helpers/sequence')
 const buildHelpers = require('../helpers/build')
 const distHelpers = require('../helpers/dist')
-const checkForAppPath = require('../helpers/checkForAppPath')
 
 module.exports = types => {
   const baseDistDir = path.join(process.cwd(), process.env.LNG_DIST_FOLDER || 'dist')
-  checkForAppPath()
 
   let metadata
 
   const dist = type => {
     let distDir
     return sequence([
+      () => buildHelpers.ensureLightningApp(),
       () => distHelpers.moveOldDistFolderToBuildFolder(),
       () => buildHelpers.ensureCorrectGitIgnore(),
       () => buildHelpers.readMetadata().then(result => (metadata = result)),
