+++ src/actions/serve.js
@@ -23,32 +23,40 @@ const chalk = require('chalk')
 const os = require('os')
 const child_process = require('child_process')
 const isLocallyInstalled = require('../helpers/localinstallationcheck')
-const checkForAppPath = require('../helpers/checkForAppPath')
+const buildHelpers = require('../helpers/build')
+const sequence = require('../helpers/sequence')
 
 module.exports = () => {
-  checkForAppPath()
-  const args = [
-    './build',
-    process.env.LNG_SERVE_OPEN === 'false' ? false : '-o',
-    process.env.LNG_SERVE_CACHE_TIME ? '-c' + process.env.LNG_SERVE_CACHE_TIME : '-c-1',
-    process.env.LNG_SERVE_PORT ? '-p' + process.env.LNG_SERVE_PORT : false,
-    process.env.LNG_SERVE_PROXY ? '-P' + process.env.LNG_SERVE_PROXY : false,
-  ].filter(val => val)
+  return sequence([
+    () => buildHelpers.ensureLightningApp(),
+    () => {
+      const args = [
+        './build',
+        process.env.LNG_SERVE_OPEN === 'false' ? false : '-o',
+        process.env.LNG_SERVE_CACHE_TIME ? '-c' + process.env.LNG_SERVE_CACHE_TIME : '-c-1',
+        process.env.LNG_SERVE_PORT ? '-p' + process.env.LNG_SERVE_PORT : false,
+        process.env.LNG_SERVE_PROXY ? '-P' + process.env.LNG_SERVE_PROXY : false,
+      ].filter(val => val)
 
-  const levelsDown = isLocallyInstalled() ? '../../../../..' : '../..'
-  const subprocess = execa(path.join(__dirname, levelsDown, 'node_modules/.bin/http-server'), args)
+      const levelsDown = isLocallyInstalled() ? '../../../../..' : '../..'
+      const subprocess = execa(
+        path.join(__dirname, levelsDown, 'node_modules/.bin/http-server'),
+        args
+      )
 
-  subprocess.catch(e => console.log(chalk.red(e.stderr)))
-  subprocess.stdout.pipe(process.stdout)
+      subprocess.catch(e => console.log(chalk.red(e.stderr)))
+      subprocess.stdout.pipe(process.stdout)
 
-  // Hack for windows to prevent leaving orphan processes, resulting in multiple http-server running instances
-  if (os.platform() === 'win32') {
-    process.on('SIGINT', () => {
-      child_process.exec('taskkill /pid ' + subprocess.pid + ' /t /f', () => {
-        process.exit()
-      })
-    })
-  }
+      // Hack for windows to prevent leaving orphan processes, resulting in multiple http-server running instances
+      if (os.platform() === 'win32') {
+        process.on('SIGINT', () => {
+          child_process.exec('taskkill /pid ' + subprocess.pid + ' /t /f', () => {
+            process.exit()
+          })
+        })
+      }
 
-  return subprocess
+      return subprocess
+    },
+  ])
 }
