+++ src/actions/upload.js
@@ -27,7 +27,6 @@ const ask = require('../helpers/ask')
 const spinner = require('../helpers/spinner')
 const exit = require('../helpers/exit')
 const buildHelpers = require('../helpers/build')
-const checkForAppPath = require('../helpers/checkForAppPath')
 
 const UPLOAD_ERRORS = {
   version_already_exists: 'The current version of your app already exists',
@@ -89,9 +88,9 @@ const upload = (packageData, user) => {
 }
 
 module.exports = () => {
-  checkForAppPath()
   let user
   return sequence([
+    () => buildHelpers.ensureLightningApp(),
     () => buildHelpers.ensureCorrectGitIgnore(),
     // todo: save API key locally for future use and set it as default answer
     () => ask('Please provide your API key'),
