+++ src/actions/watch.js
@@ -22,7 +22,6 @@ const watch = require('watch')
 const exit = require('../helpers/exit')
 const WebSocket = require('ws')
 const chalk = require('chalk')
-const checkForAppPath = require('../helpers/checkForAppPath')
 
 const regexp = /^(?!src|static|settings\.json|metadata\.json)(.+)$/
 
@@ -48,7 +47,6 @@ const initWebSocketServer = () => {
 }
 
 module.exports = (initCallback, watchCallback) => {
-  checkForAppPath()
   let busy = false
   return watch.watchTree(
     './',
