--- src/helpers/checkForAppPath.js
@@ -1,41 +1,0 @@
-/*
- * If not stated otherwise in this file or this component's LICENSE file the
- * following copyright and licenses apply:
- *
- * Copyright 2020 RDK Management
- *
- * Licensed under the Apache License, Version 2.0 (the License);
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-const exit = require('./exit')
-const fs = require('fs')
-const path = require('path')
-
-const checkForApp = () => {
-  const packageJsonPath = path.join(process.cwd(), 'package.json')
-  if (!fs.existsSync(packageJsonPath)) {
-    exit(`Package.json is not available at ${process.cwd()}. Build process cannot be proceeded`)
-  }
-  const packageJson = require(packageJsonPath)
-  if (
-    packageJson.dependencies &&
-    (Object.keys(packageJson.dependencies).indexOf('wpe-lightning-sdk') > -1 ||
-      Object.keys(packageJson.dependencies).indexOf('@lightningjs/sdk') > -1)
-  ) {
-    return true
-  } else {
-    exit('Please make sure you are running the command in the Application directory')
-  }
-}
-
-module.exports = checkForApp
